services/pyproject.toml
[project]
name = "docu-store"
version = "0.1.2"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sentinel-auth-sdk",
    "confluent-kafka>=2.13.0",
    "dateparser>=1.2.0",
    "eventsourcing>=9.5.2",
    "gliner2>=1.2.0",
    "eventsourcing-kurrentdb>=1.2.2",
    "fastapi>=0.128.0",
    "fsspec>=2026.1.0",
    "lagom>=2.7.7",
    "langchain>=1.2.9",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.9",
    "langchain-ollama>=1.0.1",
    "langchain-text-splitters>=0.3.0",
    "langfuse>=3.14.5",
    "motor>=3.5.0",
    "pdf2image>=1.17.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pymongo>=4.7.0",
    "pymupdf>=1.26.7",
    "pypdf>=6.6.2",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "pyyaml>=6.0.3",
    "qdrant-client>=1.12.0",
    "rdkit>=2025.9.5",
    "returns>=0.26.0",
    "sentence-transformers>=3.3.0",
    "structflo-cser>=0.2.0",
    "structflo-ner>=0.3.0",
    "structlog>=25.5.0",
    "temporalio>=1.3.0",
    "torch>=2.5.0",
    "transformers>=5.1.0",
    "uvicorn>=0.40.0",
    "einops",
]


[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "ALL",    # all rules
]
ignore = [
    "D100",   # Missing docstring in public module
    "D101",   # Missing docstring in public class
    "D102",   # Missing docstring in public method
    "D103",   # Missing docstring in public function
    "D104",   # Missing docstring in public package
    "D105",   # Missing docstring in magic method
    "D106",   # Missing docstring in public nested class
    "D107",   # Missing docstring in __init__
    "D417",   # Missing argument descriptions in docstring
    "E501",   # line too long (handled by formatter)
    "TC001",  # Move application import into TYPE_CHECKING block
    "TC002",  # Move third-party import into TYPE_CHECKING block
    "TC003",  # Move stdlib import into TYPE_CHECKING block
    "PLR0913",# Too many arguments in function definition
    "PLR0912",# Too many branches
    "PLR0915",# Too many statements
    "PLR2004",# Magic value used in comparison
    "C901",   # Too complex
    "BLE001", # Do not catch blind exception
    "FBT001", # Boolean-typed positional argument
    "FBT002", # Boolean default positional argument
    "ARG002", # Unused method argument
    "FAST002",# FastAPI dependency without Annotated
    "PLC0415",# Import not at top of file
    "TRY300", # Use else block rather than try-except
    "SLF001", # Private member accessed
    "S106",   # Possible hardcoded password
    "PLW2901",# for loop variable overwritten
    "N806",   # Variable in function should be lowercase
    "ANN",    # Missing type annotations (whole category)
    "COM812", # Missing trailing comma (conflicts with formatter)
    "PERF401",# Use list comprehension (style preference)
    "PERF403",# Use dict comprehension (style preference)
    "PERF102",# Use dict.values() (style preference)
    "E731",   # Do not assign lambda expression
    "SIM108", # Use ternary operator
    "SIM103", # Return negated condition directly
    "PIE810", # Call startswith once with tuple
    "D401",   # First line of docstring should be imperative mood
    "D205",   # 1 blank line required between summary and description
    "ERA001", # Found commented-out code
    "FBT003", # Boolean positional value in function call
    "C408",   # Unnecessary dict() call
    "RUF001", # String contains ambiguous unicode character
    "PYI034", # __enter__ methods usually return self
    "DTZ901", # datetime.min without timezone
    "B905",   # zip() without strict parameter
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.pylint]
disable = [
    "missing-module-docstring",
    "missing-class-docstring",
    "missing-function-docstring",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.14.14",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
markers = [
    "asyncio: marks tests as async (deselect with '-m \"not asyncio\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "slow: marks tests as slow running",
]
addopts = "--verbose --strict-markers --tb=short"
minversion = "7.0"

[tool.mypy]
python_version = "3.12"
explicit_package_bases = true
mypy_path = ["."]

